chore: move cache config to separated file
